feat: add amplitude lib
